refactor: extract tile-dialog and tile utils from group-dashboard
group-dashboard.ts had grown past 1800 lines. Pull out two clean,
gridstack-independent seams:

- dashboard-tile-dialog.ts: the add/edit-tile form as a light-DOM component.
  It owns its own draft state and URL validation and communicates with the
  parent via tile-confirmed / dialog-closed events. Light DOM so it keeps
  sharing group-dashboard's injected .add-tile-* / moss-* styles.
- dashboard-tile-utils.ts: pure helpers (NewTileDraft, DEFAULT_LAYOUTS,
  tileKindLabel, color cycle + tileColorStyle/nextTileColor, canFillHeight),
  now unit tested (yarn test:unit).

group-dashboard.ts drops ~260 lines (1812 -> 1553). The stateful gridstack
lifecycle, palette drag-in, and fill/auto-scroll machinery stay in the
component for now — extracting those safely needs runtime drag/resize
verification that the current test suite doesn't provide.

No behavior change.
